StyleGenerator: when told to export an image, always write filePath to the config file
Currently we only write 'filePath' to the config file for
images that are actually exported. For hidden images, we don't.
The result is that we need to check in all the QML file if an
image is available before we try to load it.

This patch will remove this need by simply always write a
filePath property to the config. If no image was exported, then
the path will be empty, which is effectively how all the
bindings are working already.

Change-Id: I2dded58af572789e3687635eef329c741be9b701
Reviewed-by: Doris Verria <doris.verria@qt.io>