QQmlRefCount: break all users that attempt to access the dtor
... by making the dtor private, accessible only via an auxiliary
friend class QQmlRefCounted<T>.

There are no users of QQmlRefCount outside of qtdeclarative, but we
cannot rule out that users use this private API. Since we develop this
patch as part of our own due diligence, we might as well put it into
the 6.6 release so as to alert any potential users outside the qt5.git
tree.

Once this change is released as part of 6.6, we know we have no
users of ~QQmlRefCount() (anymore), and can, in the next step, drop
the virtual from the dtor, saving sizeof(void*) on each object in the
hierarchy that isn't virtual of its own volition.

Task-number: QTBUG-114817
Change-Id: I16ea02dc57f9555e302d9415f3e573d8765cf7a3
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit f8c9fd6bfa5e6de4ac5d0176ebbc0292dbd031a2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>