qmlls: find usages of signals and bindings
Implement support to find usages of signals defined in QML or CPP and to
find usages of bindings, for the "normal" cases.
"Advanced" cases, which is about using QML objects like Connection or
Binding, will be covered in a follow-up commit.

Finding usages of signals:
Set the semantic scope of a signal's owner inside the signal, to be able
to differentiate between same-name signals of different types during
find usages. This requires to move code from
QQmlDomAstCreator::{endV,v}isit(AST::UiSourceElement*) to
{endV,v}isit(AST::FunctionDeclaration*), such that
QQmlDomAstCreatorWithQQmlJSScope can set the semantic scope in the
signal's MethodInfo during endVisit(AST::FunctionDeclaration*) instead
of endVisit(AST::UiSourceElement*) where the MethodInfo is already not
available on the node stack anymore.
Also implement finding onSignalHandler for signals as part of the
usages, and finding signals when finding usages of the onSignalHandler.

Finding usages of bindings:
Add an "identifier" region to FileLocations of bindings and property
definitions to be able to find the exact location of the property name
inside the property binding/definition. Change the existing tests to
only expect the source location of the property name instead of the
entire property definition.

Add tests for finding usages of signals and bindings and make sure that
functions contain a qqmljsscope in tst_qmldomitem::callExpressions()
and that Qml Object id usages can be found.

Disable a warning about prototypes not being a QmlObject: it is fine if
the prototype is also a QmlComponent.

Task-number: QTBUG-111414
Change-Id: Id6fd04ee7c04293a259588165f13b5ae8c78e9dc
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>