QmlCompiler: Check all names against the type location map
Value type names are lower case. If a value type from a module is used,
that also counts as usage.

The whole affair is a heuristic with or without this. An identifier can
be a number of other things that we cannot find in the imports.

Change-Id: I55a5d2dfa439d20dcbd9d015cbc2f2efbfaa8b73
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 1d12a1e93add3a1ff303e6126fdaa8cf6f5e00af)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>