Make QQmlGuard hierarchy a bit safer
Mark ~QQmlGuardImpl() as protected and QQmlStrongJSQObjectReference as
final.

QQmlGuard itself cannot be marked as final, as there are half a dozen
or so classes that use is-a instead of has-a QQmlGuard for some
reason.

Add a comment suggesting to make all the QQmlGuardImpl members
private at some point in the future.

Pick-to: 6.6 6.5
Change-Id: I66fbb80b0f8a27c963e4d520e7f1ae0588e56db4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>