Add test case for reparenting item while handler holds grab
This is a partial port of 7cf29f3250cb48842eba25815ce46ca8065f8e6e to
Qt 6.

The crash that the change fixes in Qt 5.15 doesn't happen in Qt 6, make
sure that we don't introduce issues later. We don't get grabChanged
signals emitted when the item gets reparented, so it's likely that the
handler still holds the grab, it just doesn't cause any problems for
this particular scenario.

The test counting signal emissions is expected to fail in Qt 6, we can't
clear up the grab as we don't have a currentEvent in the handler at the
point of item scene change anymore.

Task-number: QTBUG-114475
Pick-to: 6.6 6.5
Change-Id: I1cc39174c4b9b5ed1d5868610d6eacf0151cff5f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>