PointerHandler: cancel all grabs when item changes scene
When an item that has a pointer handler is taken out from a scene, then
all grabs that the handler might hold need to be cancelled. Otherwise,
the event delivery code might access data structures that are no longer
valid, like the window of the item that the grabbing handler belongs to.

The item might be removed from a scene at any point, and we can only
cancel the grab of each handler that has a current event stored, which
is the last event the handler received. If there is such an event,
then we can cancel the grab of the handler for all event points.

If the item is removed to the scene before the handler gets the event,
then the event should not get delivered to the handler at all.

Add a test that exercises the different scenarios. The test crashes
in both cases without this change.

Fixes: QTBUG-114475
Change-Id: I1022f078b02d778d8d859977c7075a4c14c9e849
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>