qmlls: find usages of propertyChanged signals/handlers
If onSignalHandler is a usage of the corresponding signal, then this
should also be true for property changed signals/handlers. For some
property p, also find onPChanged and pChanged usages (and vice-versa).

Add a heuristic to discern property changed handlers from signals from
property changed signals from etc, which will be also needed for the
renaming operation in qmlls.

Fixes: QTBUG-111414
Change-Id: Idc179a288ef2771c094efadacdeb2a5b5bb89336
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>