QQuickAbstractButton: Don't allow unchecking buttons in exclusive ActionGroups
Without this patch, buttons types that DID NOT override nextCheckState
method (like RadioButton does) failed to remain checked with an
exclusive ActionGroup bound to their actions. The code here only
accounted for ButtonGroup data. To fix this, add a condition to allow
unchecking the button only if another action is currently checked in
its exclusive ActionGroup (if any).

The whole condition was hard to interpret in its one-liner form, so I
refactored+inverted it into a tree of ifs with early returns.

Test for controls is largely copied from ButtonGroup tests. Another one
replicates the problem we are facing with drop-down menus in Plasma.

Pick-to: 6.5 6.6
Change-Id: I5bc1b4e8d9fc325a5d7e7f363975d947e5e09ace
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>