StyleGenerator: fix crash in the duplicate checking
The current implementation would access the first element
in the list of results without checking if was empty
first. This caused a crash.

Since the check for duplicates makes sense from other
places as well, move the implementation into
jsontools for everyone to enjoy, which will also
fix the crash.

Change-Id: I1c6e0d960885b16f183cc86c27e068e2608d9a1e
Reviewed-by: Doris Verria <doris.verria@qt.io>