StyleGenerator: implement 'page' property
It can easily happen that the Figma design ends up
containing several component sets with the same name.
To limit the risk of picking up the wrong one from the
style generator (which can be checked with the --sanity
switch), add the possibility to specify which
page to search inside from config.json.

Since our own templates are now all located inside the
"library" page, we can also utilize this further, and
cache the json object pointing to the page so we don't
need to search for it again and again for every control.
This addition should therefore also speed up the style
generator a bit.

Change-Id: Ibe13c9a1d9f79daf9f75f8cb0f2f74414d9cb849
Reviewed-by: Doris Verria <doris.verria@qt.io>