Fix scroll wheel issue for top item with overlay as parent
The overlay filters wheel event for the popup item as part of patch set
caca7d7d4fcb7281332904c01950a66385bad0b3. These events are consumed by
modal popup when the top level item is not its child. In a case, where
top level item is a child of overlay and stacked higher than popup, the
wheel event should be delivered to the top level item.

This patch set validates item at the top is child of overlay or popup
item and deliver wheel events accordingly.

Fixes: QTBUG-113842
Pick-to: 6.6 6.5
Change-Id: I2d27c09738742ef5ae25e42eca3747f1631baa6c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>