Make key event forwarding functional with QQuickRenderControl
...in applications that only use the public API and so reimplement
the QQuickRenderControl::renderWindowFor() virtual.

For more complex cases one needs to reimplement a virtual in the
private class as well (QQuickRenderControlPrivate::isRenderWindow())
but simple apps do not need that (e.g. because they have a single
on-screen window and a single off-screen window).

However, right now isRenderWindow() is the only function that is
used when deciding if the "window" has focus. So apps that do not
reimplement the private virtual are in trouble.

By providing a simple default implementation that falls back on
renderWindowFor (the public virtual), we can keep most
applications working and avoid the regression we've been seeing in
the past few releases.

Whereas QQuickWidget is not affected since that reimplements both the
public and private virtuals. That is an example of a complex case where
(to support graphics view proxy widgets) one needs to provide a proper,
custom implementation of the private virtual (isRenderWindow).

Amends 8c0b1e06d9679676d12ff92db981198077eeda76

Pick-to: 6.6 6.5
Fixes: QTBUG-114961
Change-Id: I024f68da7f4260191b67be89b8957b47004e8339
Reviewed-by: Andy Nichols <andy.nichols@qt.io>