StyleGenerator: enhance 'default export'
We end up repeating our selves on how we should export
the various atoms in each control. In order to make it
easier to change how we e.g export the 'background' atom
later, enhance the "default export" property to take a
list where we can specify how each atom type should be exported
by default. As such, if an atom doesn't have an explicit
'export' property, it will now use the default export that
maps to the atom name at the top of the config file.

Change-Id: Iff146cb0e2d3c3193858631b46f840a564a3300c
Reviewed-by: Doris Verria <doris.verria@qt.io>