Recycle buffer objects based on size
When we recycle batches, there is no logic to make sure we pick
a batch which is the appropriate size for the actual vertex buffer
or index buffer we will upload. In fact, when picking a batch
from the pool, we do not even know the size yet, since we have not
done any merging yet. In some cases, this can cause memory to grow
much larger than needed, in particular if we have a single large
batch and thousands of smaller ones which are updated regularly.
As we recycle "random" batches from the pool, we will gradually
resize all of them to fit the largest batch, and end up using
largest batch size * N for memory, when we could have just reused
the buffers for the largest batch in each update.

One way to fix this is to keep a separate buffer pool instead of
tying these to batches and search for the smallest buffer that can
fit the data. This means another search in a tight loop, but the
number of buffers will typically not be large, and we have such
searches in other places in the loop already, so complexity
remains the same at least.

Other ways to mitigate this might be to set a max size on the batch
pool, but this would probably require a lot of effort to find what
the sweet spots are, and in the end it's likely to be different
for different target systems.

Fixes: QTBUG-107214
Change-Id: Ie60d158348fdbda2fb79eb8bf3b61311cc67ce67
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>