StyleGenerator: improve how we resolve atom path and visibility
When we use JsonTools to search for atoms, we currently only
check the visibility of the children listed in the "figmaPath"
in the config file to figure out if an atom is effectively
visible. This is not optimal, since the designer can also
easily hide children in-between those children, which would
also effectivly hide the atom.

This patch will therefore check the visibility of all the
children along the path from the root to determine if an
atom ends up visible or not.

In order to do this correctly, we also need to change how we
resolve the path, so that it contains _all_ the children from
the root to the atom, and not only the children listed in the
"figmaPath" property. Each child in the path will further be
tagged with "[hidden]" if it was found to be hidden. This also
gives us a great way to debug why an atom might end up as
hidden by the generator, by just printing out the resolved path.

Thirdly, we place the resolved path into a map that maps
figmaId to path, instead of writing 'qt_path' and
'qt_resolvedVisible' into the found object. Doing so will
likely reduce the number of copies taken of the json object
while searching.

Change-Id: I7f472e62b1b6d27a50a5af8376e6162e5299383f
Reviewed-by: Doris Verria <doris.verria@qt.io>