qmlls: preparation to check names during rename
This commit does not add any new functionality but refactors the code in
qmllsutils to be ready for the "check if names are valid during a
rename" feature of renaming usages in qmlls.

Encapsulate the result of resolveExpressionType into its own struct.
Like this, resolveExpressionType can return additional information like
the type of identifier (a signal, a signal handler, ...) that was
resolved. Adapt the code using resolveExpressionType to use the new struct.

Rename namingAlternativesOf to resolveNameInQmlScope, as it actually
does resolve a name to whatever it represents in a QML scope (a
property, a property changed handler, a signal, and so on).

Rename findMethodIn and findPropertyIn to methodFromReferrerScope and
propertyFromReferrerScope to differentiate between the method
"hasMethodOrSignal" that "finds a method/property inside one QML scope"
and "methodFromReferrerScope" that "finds a method/property one can
call/use from the given scope that may belong to another QML Object".

Move some code from resolveExpressionType into its own static method
resolveIdentifierExpressionType. Also remove some code duplication from
resolveExpressionType and dead code (e.g. the QQmlLSUtilsResolveOptions
is only required when working on identifier expressions for qualified
names and properties).

Change-Id: I313711015675c23b53bb426a7124b8f4b57f4545
Task-number: QTBUG-114788
Task-number: QTBUG-114951
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>