QQmlSA: Remove Element ctor taking a name, add resolveBuiltinType
The Element constructor taking a name would internally create a
QQmlJSScope with a matching internal name – without doing any validation
that such an internal name would be sensible. Additionally, you could
create an Element with it, but you coudln't do anything sensible with it
as Element only has a read-only API, and the constructed QQmlJSScope
only contains a name.
Lastly, we do not really want to expose the internalName as anything
more than a transparent id at most.

Checking users of the constructor, one quickly finds that the only usage
was inside the quick plugin, which needed it to get access to the
"function" type; it also relied on "inherits" only checking the
internalName.
As an alternative, we can provide a resolveBuiltinType which can be used
for further checking, and actually returns the correct type with all its
methods and attributens, instead of a mock type which happens to work in
equality checks.

Change-Id: I212532053d1b5c898776a564f2011ba17b074079
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 255ba76c52e76dd29ab73b63c16c73c697e56bb3)