Pass render mode correctly to material for all cases
One of the code paths that call createShader() was not passing along
the render mode, causing it to default to 2D rendering. This could
cause issues when rendering a 2D item in 3D if the material has
special handling of perspective transforms, like e.g. with text
which needs a different shader to get correct antialiasing in 3D.

Pick-to: 6.2 6.5 6.6
Change-Id: I4a22642863df58351daeec63e99b54a986b4b45e
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>