QQmlDelegateModelPrivate:object: release objects only with objectRef >0
If a cache item has an object and there are no pending incubation
tasks, it returns the object.
Otherwise, it releases the object, before deleting the cache item, if
it is unreferenced.

If the cache item's object has a zero objectRef, because it is still
under construction, the method attempts to release it.
Releasing an object with a zero objectRef triggers an assertion in
QQmlDelegateModelItem::releaseObject().

The cacheItem's object is referenced in qqmldelegatemodel.cpp:1234,
which increases the objectRef from 0 to 1. It is set to 0 again,
if the cacheItem has a delegate, which gets incubated in
qqmldelegatemodel.cpp:1281. This case is not reflected later, and the
cacheItem's object is unconditionally released.

This patch makes the release attempt conditional to a positive
objectRef. It adds an autotest in tst_QQmlDelegateModel.

Fixes: QTBUG-104469
Change-Id: I5c0751ca7f796a9701889520c526f719a27a200b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit e9f631b09265f575575038f6979443143e2121c0)