StyleGenerator: add possibility to run generator in GUI mode
This patch will add support for running the StyleGenerator
with a QML GUI. It will automatically start up in GUI mode
if no fileID was provided from the command line.

In order to do that, the generator have to run in a
separate thread to not block the GUI while e.g downloading
images and such. Since the generator should no longer write
output directly to the command line, but optionally to a GUI
that runs in a separate thread, the output is now emitted as
(queued) signals.

In addition, a Bridge class has been added to transfer
generator information between QML and c++.

The generator can still be run from the command line, and
should work as before. The command line options will also
work when running in GUI mode, so that e.g --verbose and
--generate works as expected.

Change-Id: I15f83413fdf1d2ebbcb317aff5f8caa4cdebd8d1
Reviewed-by: Doris Verria <doris.verria@qt.io>