PassManager: Remove setter for category enabled
There's a use case for querying whether a logging category is enabled -
a plugin might not want to register a pass at all if it detects that a
category is not enabled.
The PassManager should however not be allowed to en- or disable
categories by itself - that should solely be handled by qmllint (or the
LSP) based on the configuration and command line arguments.

Additional clean-up: Remove the function unused and unimplemented
function prototypes in PassManagerPrivate.

Change-Id: I2c9354e6ad5988070390279b217bab862b806d82
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit fcb6bb1979459360cb885c72400a33031bfd59a8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>