QtQml: Clean up lazy-loading of parts of QQmlTypePrivate
The meta objects and the enums are lazily loaded semi-independently from
each other. Their data structures take up a lot of space. We only want
to incur that cost if they are actually needed. Furthermore, the
lazy-loading needs to be thread-safe. The easiest way to make a thread
safe lazy loading structure is via a mutable atomic pointer to an
immutable object. Refactor QQmlTypePrivate to do just that.

This reduces the memory usage of the base structure by about 50%, cuts
down on the locking required to set up the extra aspects, and generally
makes the code more readable.

Change-Id: I815338d3e4de02b2f535ab2fee748bba47719efc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>