Add explicit option for landscape mode in gallery example
The patch set 730c1d16301ba5b3a27700741b7092290ca03279 adds support for
landscape mode in gallery example. But this causes glitch when window
resizes dynamically (as in this scenario, the virtual keyboard shows up
and it dynamically resizes window).

This patch adds an option in settings to enable landscape mode only
when required and disabled by default.

Fixes: QTBUG-114693
Pick-to: 6.5 6.6
Change-Id: I64cd4d48ade10a4f43615663a0b7aef3e13a2895
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>