CMake: Duplicate backing library metatypes to its associated plugin
qmltyperegistar and other tooling expect a list of foreign metatypes
derived from the full set of dependencies of the target it's
processing.

If a target was linking to a qml plugin, and not its backing library,
the metatypes of the backing library were not passed to the tooling.

The mechanism by which the metatypes are collected is via CMake's
INTERFACE_SOURCES, and because the qml plugin links PRIVATEly to its
backing library, in the chain app <- plugin <- backing lib
the backing lib's interface sources were not passed to app.

To work around this, use new qtbase api to duplicate the metatypes
of the backing library to its associated qml plugin's
INTERFACE_SOURCES, thus avoiding the issue.

This is the best we can do given the CMake limitation of not being able
to transitively propagate values of any custom property and without
relying on finalizers.
See https://gitlab.kitware.com/cmake/cmake/-/issues/20416

Fixes: QTBUG-115152
Change-Id: I8a1652a9b12bad01abce38bb6f381949eaa1feab
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>