QQmlTreeModelToTableModel: reduce unneeded relocations
The connections array was marked as static, which forces the compiler
to place it as written, as a table of structs of pointers.

Removing the static keyword allows the compiler to generate the code
anyway it wants, incl. unrolling the loop over the data structure.

Removes 22 unneeded relocations, and saves 240B in Text and Data
combined.

Task-number: QTBUG-100536
Change-Id: I9bddd4ae4053760352a5fa359dc9336b6bd90892
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 87e3099812b4dd8c19ed022d8983d468b330471f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>