QQmlTreeModelToTableModel: reduce unneeded relocations
The connections array was marked as static, which forces the compiler
to place it as written, as a table of structs of pointers.

Removing the static keyword allows the compiler to generate the code
anyway it wants, incl. unrolling the loop over the data structure.

Removes 22 unneeded relocations, and saves 240B in Text and Data
combined.

Pick-to: 6.6
Task-number: QTBUG-100536
Change-Id: I9bddd4ae4053760352a5fa359dc9336b6bd90892
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>