StyleGenerator Slider: Use the size of the groove if background is not visible
With the current implementation where the implicit size of the
background item, was that of the background image, we failed to show
the groove and track inside, if the background was set to hidden in
Figma.
To fix, if the background image is not visible, set the implicit size
of the background item, to the groove's implicit size.

Change-Id: Ia7f9fe87834e017b39d5ede04f2633d9107c61dd
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>