Remove bounding behavior of popup for negative margins
The popup behavior is inconsistent when its moved or repositioned over
the edges of the bounding window. This happens for negative margins,
as its made to bound within the window (while respositioning using
QQuickPopupPositioner::reposition) only on left and top edges of
the window during resize, but its allowed to move outside the bounding
window on right and bottom edges. This pushing of popup within the bound
can also sometimes leads to infinite polish issues (for the qml
configuration as mentioned in the bug report QTBUG-107473).

This patch set removes this constraints for popup to move beyond edges of
the window (through QQuickPopupPrivate::relaxEdgeConstraint) during
resize. By default, this flag is disabled, means those controls which
inherits from popup restricts by itself not to move beyond the edges
during resize.

[ChangeLog][Qt Quick Controls][Popup][Important Behavior Changes] The popup
default margins (-1) earlier was constrained on left and top edges, but
allowed to move beyond right and bottom edges. The behavior changed now to
relax this constraint and move beyond all edges. Also to be noted, edge
constraints has been relaxed only for Popup and controls inheriting from
popup (such as drawer, menu) still follows earlier behavior.

Fixes: QTBUG-107473
Fixes: QTBUG-77647
Change-Id: Icf0dc909559b5059a7371a6455a15ffc84e9b77d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>