StyleGenerator, GUI: enable filtering of output
By using a filter, the user can more easily search for
output related to e.g a single control. We can then also
store all output, and ignore the "verbose" switch.
Adding in addition a checkbox that let's the user show or
hide the log at will. But the log will still automatically
show in case of warnings.

Change-Id: I1346e95ca6824f5e8c4d62b2d503cbb18a4f4c3b
Reviewed-by: Doris Verria <doris.verria@qt.io>